backend/requirements.txt
fastapi 
uvicorn 
requests 
cloudscraper 
beautifulsoup4 
python-dotenv 
groq 
pydantic 
urllib3 
